// client/src/store/messages/actionTypes.ts
export enum ActionTypes {
  FETCH_MESSAGES_START = "MESSAGES:FETCH_MESSAGES_START",
  FETCH_MESSAGES_FAIL = "MESSAGES:FETCH_MESSAGES_FAIL",
  FETCH_MESSAGES_SUCCESS = "MESSAGES:FETCH_MESSAGES_SUCCESS",
  ADD_MESSAGE = "MESSAGES:ADD_MESSAGE",
  RESET_MESSAGES = "MESSAGES:RESET_MESSAGES",
  POST_MESSAGE_START = "MESSAGES:POST_MESSAGE_START",
  POST_MESSAGE_FAIL = "MESSAGES:POST_MESSAGE_FAIL",
  POST_DELETE_MESSAGE_START = "MESSAGES:POST_DELETE_MESSAGE_START",
  POST_DELETE_MESSAGE_FAIL = "MESSAGES:POST_DELETE_MESSAGE_FAIL",
  REMOVE_MESSAGE = "MESSAGES:REMOVE_MESSAGE",
}